+++ app/Http/Controllers/DiscountController.php
@@ -4,9 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\Discount\StoreDiscountRequest;
 use App\Http\Requests\Discount\UpdateDiscountRequest;
+use App\Http\Resources\Discount\DiscountCollection;
 use App\Http\Resources\Discount\DiscountResource;
 use App\Models\Discount;
+use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Response as HttpResponse;
 
 class DiscountController extends Controller
@@ -21,11 +24,8 @@ class DiscountController extends Controller
      */
     public function index()
     {
-        $discount = $this->discount->paginate(5);
-        $discountResource = DiscountResource::collection($discount)->response()->getData(true);
-        return response()->json([
-            'data' => $discountResource,
-        ], HttpResponse::HTTP_OK);
+        $discountResource = new DiscountCollection(Discount::all());
+        return $discountResource;
     }
 
     /**
@@ -34,16 +34,22 @@ class DiscountController extends Controller
     public function store(StoreDiscountRequest $request)
     {
         $dataCreate = $request->all();
-        $check = Discount::where('discount', $request->discount)->first();
-        if($check){
-            flash()->addError('Mã giảm giá này đã tồn tại');
-            return redirect()->back();
+        $check = Discount::where('discount', $dataCreate['discount'])->exists();
+        if ($check) {
+            return response()->json([
+                'error' => 'Discount này đã tồn tại!'
+            ], HttpResponse::HTTP_CONFLICT);
         }
-        $discount = $this->discount->create($dataCreate);
-        $discountResource = new DiscountResource($discount);
-        return response()->json([
-            'data' => $discountResource,
-        ], HttpResponse::HTTP_OK);
+        $discount = new Discount();
+        $discount->discount = $dataCreate['discount'];
+        $discount->start_day = $dataCreate['start_day'];
+        $discount->end_day = $dataCreate['end_day'];
+        $discount->status = $dataCreate['status'];
+        $discount->note = $dataCreate['note'];
+        $discount->save();
+        return (new DiscountResource($discount))
+            ->response()
+            ->setStatusCode(HttpResponse::HTTP_OK);
     }
 
     /**
@@ -59,13 +65,29 @@ class DiscountController extends Controller
      */
     public function update(UpdateDiscountRequest $request, string $id)
     {
-        $discount = $this->discount->findOrFail($id);
-        $dataUpdate = $request->all();
-        $discount->update($dataUpdate);
-        $discountResource = new DiscountResource($discount);
-        return response()->json([
-            'data' => $discountResource,
-        ], HttpResponse::HTTP_OK);
+        try {
+            $discount = $this->discount->findOrFail($id);
+            $dataUpdate = $request->all();
+            $check = Discount::where('discount', $dataUpdate['discount'])->where('discount_id', '!=', $id)->exists();
+            if ($check) {
+                return response()->json([
+                    'error' => 'Discount này đã tồn tại!',
+                ], HttpResponse::HTTP_CONFLICT);
+            }
+            $discount->discount = $dataUpdate['discount'];
+            $discount->start_day = $dataUpdate['start_day'];
+            $discount->end_day = $dataUpdate['end_day'];
+            $discount->status = $dataUpdate['status'];
+            $discount->note = $dataUpdate['note'];
+            $discount->save();
+            return (new DiscountResource($discount))
+                ->response()
+                ->setStatusCode(HttpResponse::HTTP_OK);
+        } catch (ModelNotFoundException $e) {
+            return response()->json([
+                'error' => 'Discount không tồn tại'
+            ], HttpResponse::HTTP_NOT_FOUND);
+        }
     }
 
     /**
@@ -73,11 +95,22 @@ class DiscountController extends Controller
      */
     public function destroy(string $id)
     {
-        $discount = $this->discount->where('discount_id', $id)->firstOrFail();
-        $discount->delete();
-        $discountResource = new DiscountResource($discount);
-        return response()->json([
-            'data' => $discountResource,
-        ], HttpResponse::HTTP_OK);
+        $isUsedInOtherTable = Product::where('discount_id', $id)->exists();
+        if ($isUsedInOtherTable) {
+            return response()->json([
+                'error' => 'Discount này đang có sản phẩm nên không thể xóa.',
+            ], HttpResponse::HTTP_CONFLICT);
+        }
+        try {
+            $discount = $this->discount->where('discount_id', $id)->firstOrFail();
+            $discount->delete();
+            return response()->json([
+                'message' => 'Xóa thành công ' . $discount->discount,
+            ], HttpResponse::HTTP_OK);
+        } catch (ModelNotFoundException $e) {
+            return response()->json([
+                'error' => 'Discount không tồn tại'
+            ], HttpResponse::HTTP_NOT_FOUND);
+        }
     }
 }
